@@ -1,13 +1,15 @@
 name: ory_keto_client
-version: 0.11.0-alpha.0
-description: OpenAPI API client
+version: 25.4.0
+description: OpenAPI API client for Ory Keto, Ory&#39;s open source &#39;Zanzibar - Google&#39;s Consistent, Global Authorization System&#39; implementation.
 homepage: https://www.ory.sh
+repository: https://github.com/ory/sdk/tree/master/clients/keto/dart
+
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.18.0 <4.0.0'
 
 dependencies:
-  dio: '>=4.0.1 <5.0.0'
+  dio: '^5.7.0'
   one_of: '>=1.5.0 <2.0.0'
   one_of_serializer: '>=1.5.0 <2.0.0'
   built_value: '>=8.4.0 <9.0.0'
@@ -16,4 +18,4 @@ dependencies:
 dev_dependencies:
   built_value_generator: '>=8.4.0 <9.0.0'
   build_runner: any
-  test: ^1.16.0
+  test: '^1.16.0'
