@@ -4,10 +4,13 @@
 
 import 'dart:async';
 
+import 'package:built_value/json_object.dart';
 import 'package:built_value/serializer.dart';
 import 'package:dio/dio.dart';
 
 import 'package:ory_keto_client/src/api_util.dart';
+import 'package:ory_keto_client/src/model/batch_check_permission_body.dart';
+import 'package:ory_keto_client/src/model/batch_check_permission_result.dart';
 import 'package:ory_keto_client/src/model/check_permission_result.dart';
 import 'package:ory_keto_client/src/model/error_generic.dart';
 import 'package:ory_keto_client/src/model/expanded_permission_tree.dart';
@@ -21,6 +24,109 @@ class PermissionApi {
   final Serializers _serializers;
 
   const PermissionApi(this._dio, this._serializers);
+
+  /// Batch check permissions
+  /// To learn how relationship tuples and the check works, head over to [the documentation](https://www.ory.sh/docs/keto/concepts/api-overview).
+  ///
+  /// Parameters:
+  /// * [maxDepth] 
+  /// * [batchCheckPermissionBody] 
+  /// * [cancelToken] - A [CancelToken] that can be used to cancel the operation
+  /// * [headers] - Can be used to add additional headers to the request
+  /// * [extras] - Can be used to add flags to the request
+  /// * [validateStatus] - A [ValidateStatus] callback that can be used to determine request success based on the HTTP status of the response
+  /// * [onSendProgress] - A [ProgressCallback] that can be used to get the send progress
+  /// * [onReceiveProgress] - A [ProgressCallback] that can be used to get the receive progress
+  ///
+  /// Returns a [Future] containing a [Response] with a [BatchCheckPermissionResult] as data
+  /// Throws [DioException] if API call or serialization fails
+  Future<Response<BatchCheckPermissionResult>> batchCheckPermission({ 
+    int? maxDepth,
+    BatchCheckPermissionBody? batchCheckPermissionBody,
+    CancelToken? cancelToken,
+    Map<String, dynamic>? headers,
+    Map<String, dynamic>? extra,
+    ValidateStatus? validateStatus,
+    ProgressCallback? onSendProgress,
+    ProgressCallback? onReceiveProgress,
+  }) async {
+    final _path = r'/relation-tuples/batch/check';
+    final _options = Options(
+      method: r'POST',
+      headers: <String, dynamic>{
+        ...?headers,
+      },
+      extra: <String, dynamic>{
+        'secure': <Map<String, String>>[],
+        ...?extra,
+      },
+      contentType: 'application/json',
+      validateStatus: validateStatus,
+    );
+
+    final _queryParameters = <String, dynamic>{
+      if (maxDepth != null) r'max-depth': encodeQueryParameter(_serializers, maxDepth, const FullType(int)),
+    };
+
+    dynamic _bodyData;
+
+    try {
+      const _type = FullType(BatchCheckPermissionBody);
+      _bodyData = batchCheckPermissionBody == null ? null : _serializers.serialize(batchCheckPermissionBody, specifiedType: _type);
+
+    } catch(error, stackTrace) {
+      throw DioException(
+         requestOptions: _options.compose(
+          _dio.options,
+          _path,
+          queryParameters: _queryParameters,
+        ),
+        type: DioExceptionType.unknown,
+        error: error,
+        stackTrace: stackTrace,
+      );
+    }
+
+    final _response = await _dio.request<Object>(
+      _path,
+      data: _bodyData,
+      options: _options,
+      queryParameters: _queryParameters,
+      cancelToken: cancelToken,
+      onSendProgress: onSendProgress,
+      onReceiveProgress: onReceiveProgress,
+    );
+
+    BatchCheckPermissionResult? _responseData;
+
+    try {
+      final rawResponse = _response.data;
+      _responseData = rawResponse == null ? null : _serializers.deserialize(
+        rawResponse,
+        specifiedType: const FullType(BatchCheckPermissionResult),
+      ) as BatchCheckPermissionResult;
+
+    } catch (error, stackTrace) {
+      throw DioException(
+        requestOptions: _response.requestOptions,
+        response: _response,
+        type: DioExceptionType.unknown,
+        error: error,
+        stackTrace: stackTrace,
+      );
+    }
+
+    return Response<BatchCheckPermissionResult>(
+      data: _responseData,
+      headers: _response.headers,
+      isRedirect: _response.isRedirect,
+      requestOptions: _response.requestOptions,
+      redirects: _response.redirects,
+      statusCode: _response.statusCode,
+      statusMessage: _response.statusMessage,
+      extra: _response.extra,
+    );
+  }
 
   /// Check a permission
   /// To learn how relationship tuples and the check works, head over to [the documentation](https://www.ory.sh/docs/keto/concepts/api-overview).
@@ -42,7 +148,7 @@ class PermissionApi {
   /// * [onReceiveProgress] - A [ProgressCallback] that can be used to get the receive progress
   ///
   /// Returns a [Future] containing a [Response] with a [CheckPermissionResult] as data
-  /// Throws [DioError] if API call or serialization fails
+  /// Throws [DioException] if API call or serialization fails
   Future<Response<CheckPermissionResult>> checkPermission({ 
     String? namespace,
     String? object,
@@ -92,22 +198,23 @@ class PermissionApi {
       onReceiveProgress: onReceiveProgress,
     );
 
-    CheckPermissionResult _responseData;
+    CheckPermissionResult? _responseData;
 
     try {
-      const _responseType = FullType(CheckPermissionResult);
-      _responseData = _serializers.deserialize(
-        _response.data!,
-        specifiedType: _responseType,
+      final rawResponse = _response.data;
+      _responseData = rawResponse == null ? null : _serializers.deserialize(
+        rawResponse,
+        specifiedType: const FullType(CheckPermissionResult),
       ) as CheckPermissionResult;
 
     } catch (error, stackTrace) {
-      throw DioError(
+      throw DioException(
         requestOptions: _response.requestOptions,
         response: _response,
-        type: DioErrorType.other,
+        type: DioExceptionType.unknown,
         error: error,
-      )..stackTrace = stackTrace;
+        stackTrace: stackTrace,
+      );
     }
 
     return Response<CheckPermissionResult>(
@@ -142,7 +249,7 @@ class PermissionApi {
   /// * [onReceiveProgress] - A [ProgressCallback] that can be used to get the receive progress
   ///
   /// Returns a [Future] containing a [Response] with a [CheckPermissionResult] as data
-  /// Throws [DioError] if API call or serialization fails
+  /// Throws [DioException] if API call or serialization fails
   Future<Response<CheckPermissionResult>> checkPermissionOrError({ 
     String? namespace,
     String? object,
@@ -192,22 +299,23 @@ class PermissionApi {
       onReceiveProgress: onReceiveProgress,
     );
 
-    CheckPermissionResult _responseData;
+    CheckPermissionResult? _responseData;
 
     try {
-      const _responseType = FullType(CheckPermissionResult);
-      _responseData = _serializers.deserialize(
-        _response.data!,
-        specifiedType: _responseType,
+      final rawResponse = _response.data;
+      _responseData = rawResponse == null ? null : _serializers.deserialize(
+        rawResponse,
+        specifiedType: const FullType(CheckPermissionResult),
       ) as CheckPermissionResult;
 
     } catch (error, stackTrace) {
-      throw DioError(
+      throw DioException(
         requestOptions: _response.requestOptions,
         response: _response,
-        type: DioErrorType.other,
+        type: DioExceptionType.unknown,
         error: error,
-      )..stackTrace = stackTrace;
+        stackTrace: stackTrace,
+      );
     }
 
     return Response<CheckPermissionResult>(
@@ -238,7 +346,7 @@ class PermissionApi {
   /// * [onReceiveProgress] - A [ProgressCallback] that can be used to get the receive progress
   ///
   /// Returns a [Future] containing a [Response] with a [ExpandedPermissionTree] as data
-  /// Throws [DioError] if API call or serialization fails
+  /// Throws [DioException] if API call or serialization fails
   Future<Response<ExpandedPermissionTree>> expandPermissions({ 
     required String namespace,
     required String object,
@@ -280,22 +388,23 @@ class PermissionApi {
       onReceiveProgress: onReceiveProgress,
     );
 
-    ExpandedPermissionTree _responseData;
+    ExpandedPermissionTree? _responseData;
 
     try {
-      const _responseType = FullType(ExpandedPermissionTree);
-      _responseData = _serializers.deserialize(
-        _response.data!,
-        specifiedType: _responseType,
+      final rawResponse = _response.data;
+      _responseData = rawResponse == null ? null : _serializers.deserialize(
+        rawResponse,
+        specifiedType: const FullType(ExpandedPermissionTree),
       ) as ExpandedPermissionTree;
 
     } catch (error, stackTrace) {
-      throw DioError(
+      throw DioException(
         requestOptions: _response.requestOptions,
         response: _response,
-        type: DioErrorType.other,
+        type: DioExceptionType.unknown,
         error: error,
-      )..stackTrace = stackTrace;
+        stackTrace: stackTrace,
+      );
     }
 
     return Response<ExpandedPermissionTree>(
@@ -324,7 +433,7 @@ class PermissionApi {
   /// * [onReceiveProgress] - A [ProgressCallback] that can be used to get the receive progress
   ///
   /// Returns a [Future] containing a [Response] with a [CheckPermissionResult] as data
-  /// Throws [DioError] if API call or serialization fails
+  /// Throws [DioException] if API call or serialization fails
   Future<Response<CheckPermissionResult>> postCheckPermission({ 
     int? maxDepth,
     PostCheckPermissionBody? postCheckPermissionBody,
@@ -360,15 +469,16 @@ class PermissionApi {
       _bodyData = postCheckPermissionBody == null ? null : _serializers.serialize(postCheckPermissionBody, specifiedType: _type);
 
     } catch(error, stackTrace) {
-      throw DioError(
+      throw DioException(
          requestOptions: _options.compose(
           _dio.options,
           _path,
           queryParameters: _queryParameters,
         ),
-        type: DioErrorType.other,
+        type: DioExceptionType.unknown,
         error: error,
-      )..stackTrace = stackTrace;
+        stackTrace: stackTrace,
+      );
     }
 
     final _response = await _dio.request<Object>(
@@ -381,22 +491,23 @@ class PermissionApi {
       onReceiveProgress: onReceiveProgress,
     );
 
-    CheckPermissionResult _responseData;
+    CheckPermissionResult? _responseData;
 
     try {
-      const _responseType = FullType(CheckPermissionResult);
-      _responseData = _serializers.deserialize(
-        _response.data!,
-        specifiedType: _responseType,
+      final rawResponse = _response.data;
+      _responseData = rawResponse == null ? null : _serializers.deserialize(
+        rawResponse,
+        specifiedType: const FullType(CheckPermissionResult),
       ) as CheckPermissionResult;
 
     } catch (error, stackTrace) {
-      throw DioError(
+      throw DioException(
         requestOptions: _response.requestOptions,
         response: _response,
-        type: DioErrorType.other,
+        type: DioExceptionType.unknown,
         error: error,
-      )..stackTrace = stackTrace;
+        stackTrace: stackTrace,
+      );
     }
 
     return Response<CheckPermissionResult>(
@@ -415,7 +526,7 @@ class PermissionApi {
   /// To learn how relationship tuples and the check works, head over to [the documentation](https://www.ory.sh/docs/keto/concepts/api-overview).
   ///
   /// Parameters:
-  /// * [maxDepth] - nolint:deadcode,unused
+  /// * [maxDepth] 
   /// * [postCheckPermissionOrErrorBody] 
   /// * [cancelToken] - A [CancelToken] that can be used to cancel the operation
   /// * [headers] - Can be used to add additional headers to the request
@@ -425,7 +536,7 @@ class PermissionApi {
   /// * [onReceiveProgress] - A [ProgressCallback] that can be used to get the receive progress
   ///
   /// Returns a [Future] containing a [Response] with a [CheckPermissionResult] as data
-  /// Throws [DioError] if API call or serialization fails
+  /// Throws [DioException] if API call or serialization fails
   Future<Response<CheckPermissionResult>> postCheckPermissionOrError({ 
     int? maxDepth,
     PostCheckPermissionOrErrorBody? postCheckPermissionOrErrorBody,
@@ -461,15 +572,16 @@ class PermissionApi {
       _bodyData = postCheckPermissionOrErrorBody == null ? null : _serializers.serialize(postCheckPermissionOrErrorBody, specifiedType: _type);
 
     } catch(error, stackTrace) {
-      throw DioError(
+      throw DioException(
          requestOptions: _options.compose(
           _dio.options,
           _path,
           queryParameters: _queryParameters,
         ),
-        type: DioErrorType.other,
+        type: DioExceptionType.unknown,
         error: error,
-      )..stackTrace = stackTrace;
+        stackTrace: stackTrace,
+      );
     }
 
     final _response = await _dio.request<Object>(
@@ -482,22 +594,23 @@ class PermissionApi {
       onReceiveProgress: onReceiveProgress,
     );
 
-    CheckPermissionResult _responseData;
+    CheckPermissionResult? _responseData;
 
     try {
-      const _responseType = FullType(CheckPermissionResult);
-      _responseData = _serializers.deserialize(
-        _response.data!,
-        specifiedType: _responseType,
+      final rawResponse = _response.data;
+      _responseData = rawResponse == null ? null : _serializers.deserialize(
+        rawResponse,
+        specifiedType: const FullType(CheckPermissionResult),
       ) as CheckPermissionResult;
 
     } catch (error, stackTrace) {
-      throw DioError(
+      throw DioException(
         requestOptions: _response.requestOptions,
         response: _response,
-        type: DioErrorType.other,
+        type: DioExceptionType.unknown,
         error: error,
-      )..stackTrace = stackTrace;
+        stackTrace: stackTrace,
+      );
     }
 
     return Response<CheckPermissionResult>(
