@@ -4,6 +4,7 @@
 
 import 'dart:async';
 
+import 'package:built_value/json_object.dart';
 import 'package:built_value/serializer.dart';
 import 'package:dio/dio.dart';
 
@@ -38,7 +39,7 @@ class RelationshipApi {
   /// * [onReceiveProgress] - A [ProgressCallback] that can be used to get the receive progress
   ///
   /// Returns a [Future] containing a [Response] with a [CheckOplSyntaxResult] as data
-  /// Throws [DioError] if API call or serialization fails
+  /// Throws [DioException] if API call or serialization fails
   Future<Response<CheckOplSyntaxResult>> checkOplSyntax({ 
     String? body,
     CancelToken? cancelToken,
@@ -68,14 +69,15 @@ class RelationshipApi {
       _bodyData = body;
 
     } catch(error, stackTrace) {
-      throw DioError(
+      throw DioException(
          requestOptions: _options.compose(
           _dio.options,
           _path,
         ),
-        type: DioErrorType.other,
+        type: DioExceptionType.unknown,
         error: error,
-      )..stackTrace = stackTrace;
+        stackTrace: stackTrace,
+      );
     }
 
     final _response = await _dio.request<Object>(
@@ -87,22 +89,23 @@ class RelationshipApi {
       onReceiveProgress: onReceiveProgress,
     );
 
-    CheckOplSyntaxResult _responseData;
+    CheckOplSyntaxResult? _responseData;
 
     try {
-      const _responseType = FullType(CheckOplSyntaxResult);
-      _responseData = _serializers.deserialize(
-        _response.data!,
-        specifiedType: _responseType,
+      final rawResponse = _response.data;
+      _responseData = rawResponse == null ? null : _serializers.deserialize(
+        rawResponse,
+        specifiedType: const FullType(CheckOplSyntaxResult),
       ) as CheckOplSyntaxResult;
 
     } catch (error, stackTrace) {
-      throw DioError(
+      throw DioException(
         requestOptions: _response.requestOptions,
         response: _response,
-        type: DioErrorType.other,
+        type: DioExceptionType.unknown,
         error: error,
-      )..stackTrace = stackTrace;
+        stackTrace: stackTrace,
+      );
     }
 
     return Response<CheckOplSyntaxResult>(
@@ -130,7 +133,7 @@ class RelationshipApi {
   /// * [onReceiveProgress] - A [ProgressCallback] that can be used to get the receive progress
   ///
   /// Returns a [Future] containing a [Response] with a [Relationship] as data
-  /// Throws [DioError] if API call or serialization fails
+  /// Throws [DioException] if API call or serialization fails
   Future<Response<Relationship>> createRelationship({ 
     CreateRelationshipBody? createRelationshipBody,
     CancelToken? cancelToken,
@@ -161,14 +164,15 @@ class RelationshipApi {
       _bodyData = createRelationshipBody == null ? null : _serializers.serialize(createRelationshipBody, specifiedType: _type);
 
     } catch(error, stackTrace) {
-      throw DioError(
+      throw DioException(
          requestOptions: _options.compose(
           _dio.options,
           _path,
         ),
-        type: DioErrorType.other,
+        type: DioExceptionType.unknown,
         error: error,
-      )..stackTrace = stackTrace;
+        stackTrace: stackTrace,
+      );
     }
 
     final _response = await _dio.request<Object>(
@@ -180,22 +184,23 @@ class RelationshipApi {
       onReceiveProgress: onReceiveProgress,
     );
 
-    Relationship _responseData;
+    Relationship? _responseData;
 
     try {
-      const _responseType = FullType(Relationship);
-      _responseData = _serializers.deserialize(
-        _response.data!,
-        specifiedType: _responseType,
+      final rawResponse = _response.data;
+      _responseData = rawResponse == null ? null : _serializers.deserialize(
+        rawResponse,
+        specifiedType: const FullType(Relationship),
       ) as Relationship;
 
     } catch (error, stackTrace) {
-      throw DioError(
+      throw DioException(
         requestOptions: _response.requestOptions,
         response: _response,
-        type: DioErrorType.other,
+        type: DioExceptionType.unknown,
         error: error,
-      )..stackTrace = stackTrace;
+        stackTrace: stackTrace,
+      );
     }
 
     return Response<Relationship>(
@@ -229,7 +234,7 @@ class RelationshipApi {
   /// * [onReceiveProgress] - A [ProgressCallback] that can be used to get the receive progress
   ///
   /// Returns a [Future]
-  /// Throws [DioError] if API call or serialization fails
+  /// Throws [DioException] if API call or serialization fails
   Future<Response<void>> deleteRelationships({ 
     String? namespace,
     String? object,
@@ -284,8 +289,8 @@ class RelationshipApi {
   /// Get all relationships that match the query. Only the namespace field is required.
   ///
   /// Parameters:
-  /// * [pageToken] 
-  /// * [pageSize] 
+  /// * [pageSize] - Items per Page  This is the number of items per page to return. For details on pagination please head over to the [pagination documentation](https://www.ory.sh/docs/ecosystem/api-design#pagination).
+  /// * [pageToken] - Next Page Token  The next page token. For details on pagination please head over to the [pagination documentation](https://www.ory.sh/docs/ecosystem/api-design#pagination).
   /// * [namespace] - Namespace of the Relationship
   /// * [object] - Object of the Relationship
   /// * [relation] - Relation of the Relationship
@@ -301,10 +306,10 @@ class RelationshipApi {
   /// * [onReceiveProgress] - A [ProgressCallback] that can be used to get the receive progress
   ///
   /// Returns a [Future] containing a [Response] with a [Relationships] as data
-  /// Throws [DioError] if API call or serialization fails
+  /// Throws [DioException] if API call or serialization fails
   Future<Response<Relationships>> getRelationships({ 
+    int? pageSize = 250,
     String? pageToken,
-    int? pageSize,
     String? namespace,
     String? object,
     String? relation,
@@ -333,8 +338,8 @@ class RelationshipApi {
     );
 
     final _queryParameters = <String, dynamic>{
-      if (pageToken != null) r'page_token': encodeQueryParameter(_serializers, pageToken, const FullType(String)),
       if (pageSize != null) r'page_size': encodeQueryParameter(_serializers, pageSize, const FullType(int)),
+      if (pageToken != null) r'page_token': encodeQueryParameter(_serializers, pageToken, const FullType(String)),
       if (namespace != null) r'namespace': encodeQueryParameter(_serializers, namespace, const FullType(String)),
       if (object != null) r'object': encodeQueryParameter(_serializers, object, const FullType(String)),
       if (relation != null) r'relation': encodeQueryParameter(_serializers, relation, const FullType(String)),
@@ -353,22 +358,23 @@ class RelationshipApi {
       onReceiveProgress: onReceiveProgress,
     );
 
-    Relationships _responseData;
+    Relationships? _responseData;
 
     try {
-      const _responseType = FullType(Relationships);
-      _responseData = _serializers.deserialize(
-        _response.data!,
-        specifiedType: _responseType,
+      final rawResponse = _response.data;
+      _responseData = rawResponse == null ? null : _serializers.deserialize(
+        rawResponse,
+        specifiedType: const FullType(Relationships),
       ) as Relationships;
 
     } catch (error, stackTrace) {
-      throw DioError(
+      throw DioException(
         requestOptions: _response.requestOptions,
         response: _response,
-        type: DioErrorType.other,
+        type: DioExceptionType.unknown,
         error: error,
-      )..stackTrace = stackTrace;
+        stackTrace: stackTrace,
+      );
     }
 
     return Response<Relationships>(
@@ -395,7 +401,7 @@ class RelationshipApi {
   /// * [onReceiveProgress] - A [ProgressCallback] that can be used to get the receive progress
   ///
   /// Returns a [Future] containing a [Response] with a [RelationshipNamespaces] as data
-  /// Throws [DioError] if API call or serialization fails
+  /// Throws [DioException] if API call or serialization fails
   Future<Response<RelationshipNamespaces>> listRelationshipNamespaces({ 
     CancelToken? cancelToken,
     Map<String, dynamic>? headers,
@@ -425,22 +431,23 @@ class RelationshipApi {
       onReceiveProgress: onReceiveProgress,
     );
 
-    RelationshipNamespaces _responseData;
+    RelationshipNamespaces? _responseData;
 
     try {
-      const _responseType = FullType(RelationshipNamespaces);
-      _responseData = _serializers.deserialize(
-        _response.data!,
-        specifiedType: _responseType,
+      final rawResponse = _response.data;
+      _responseData = rawResponse == null ? null : _serializers.deserialize(
+        rawResponse,
+        specifiedType: const FullType(RelationshipNamespaces),
       ) as RelationshipNamespaces;
 
     } catch (error, stackTrace) {
-      throw DioError(
+      throw DioException(
         requestOptions: _response.requestOptions,
         response: _response,
-        type: DioErrorType.other,
+        type: DioExceptionType.unknown,
         error: error,
-      )..stackTrace = stackTrace;
+        stackTrace: stackTrace,
+      );
     }
 
     return Response<RelationshipNamespaces>(
@@ -468,7 +475,7 @@ class RelationshipApi {
   /// * [onReceiveProgress] - A [ProgressCallback] that can be used to get the receive progress
   ///
   /// Returns a [Future]
-  /// Throws [DioError] if API call or serialization fails
+  /// Throws [DioException] if API call or serialization fails
   Future<Response<void>> patchRelationships({ 
     BuiltList<RelationshipPatch>? relationshipPatch,
     CancelToken? cancelToken,
@@ -499,14 +506,15 @@ class RelationshipApi {
       _bodyData = relationshipPatch == null ? null : _serializers.serialize(relationshipPatch, specifiedType: _type);
 
     } catch(error, stackTrace) {
-      throw DioError(
+      throw DioException(
          requestOptions: _options.compose(
           _dio.options,
           _path,
         ),
-        type: DioErrorType.other,
+        type: DioExceptionType.unknown,
         error: error,
-      )..stackTrace = stackTrace;
+        stackTrace: stackTrace,
+      );
     }
 
     final _response = await _dio.request<Object>(
